+++ projects/ng-components/src/lib/fhi-tab-bar/fhi-tab-bar.component.ts
@@ -1,10 +1,11 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input, ViewEncapsulation } from '@angular/core';
 
 
 @Component({
   selector: 'fhi-tab-bar',
   templateUrl: './fhi-tab-bar.component.html',
-  styleUrls: ['./fhi-tab-bar.scss']
+  styleUrls: ['./fhi-tab-bar.scss'],
+  encapsulation: ViewEncapsulation.None
 })
 export class FhiTabBarComponent {